+++ src/SwapExtractionTool/SwapTransaction.cs
@@ -2,7 +2,7 @@
 {
     public sealed class SwapTransaction
     {
-        public long BlockHeight { get; set; }
+        public int BlockHeight { get; set; }
         public string StraxAddress { get; set; }
         public decimal SenderAmount { get; set; }
         public string TransactionHash { get; set; }
